RubiconLite is now just rubicon adapter.  "rubicon" aliases "rubiconLite"
