+++ adapters.json
@@ -19,7 +19,6 @@
     "pubmatic",
     "pulsepoint",
     "rubicon",
-    "rubiconLite",
     "sonobi",
     "sovrn",
     "springserve",
@@ -51,6 +50,11 @@
     {
       "appnexusAst": {
         "supportedMediaTypes": ["video"]
+      }
+    },
+    {
+      "rubicon": {
+        "alias": "rubiconLite"
       }
     }
 ]
