+++ test/spec/adapters/rubicon_spec.js
@@ -1,47 +1,81 @@
-import { expect } from "chai";
-import adloader from "src/adloader";
-import adapterManager from "src/adaptermanager";
-import bidManager from "src/bidmanager";
-import RubiconAdapter from "src/adapters/rubicon";
+import { expect } from 'chai';
+import adloader from 'src/adloader';
+import adapterManager from 'src/adaptermanager';
+import bidManager from 'src/bidmanager';
+import RubiconAdapter from 'src/adapters/rubicon';
+import {parse as parseQuery} from 'querystring';
 
-var CONSTANTS = require("src/constants.json");
+var CONSTANTS = require('src/constants.json');
 
+describe('the rubicon adapter', () => {
 
-describe("the rubicon adapter", () => {
-
-  let rubiconAdapter,
-      sandbox,
-      adUnit;
+  let sandbox,
+      adUnit,
+      bidderRequest;
 
   beforeEach(() => {
     sandbox = sinon.sandbox.create();
 
-    delete window.rubicontag;
-
     adUnit = {
-      code: "/19968336/header-bid-tag-0",
+      code: '/19968336/header-bid-tag-0',
       sizes: [[300, 250], [320, 50]],
       bids: [
         {
-          bidder: "rubicon",
+          bidder: 'rubicon',
           params: {
-            accountId: "14062",
-            siteId: "70608",
-            zoneId: "335918",
-            userId: "12346",
-            keywords: ["a","b","c"],
+            accountId: '14062',
+            siteId: '70608',
+            zoneId: '335918',
+            userId: '12346',
+            keywords: ['a','b','c'],
             inventory: {
-              rating:"5-star",
-              prodtype:"tech"
+              rating:'5-star',
+              prodtype:'tech'
             },
             visitor: {
-              ucat:"new",
-              lastsearch:"iphone"
+              ucat:'new',
+              lastsearch:'iphone'
             },
-            position: "atf"
+            position: 'atf',
+            referrer: 'localhost'
           }
         }
       ]
+    };
+
+    bidderRequest = {
+      bidderCode: 'rubicon',
+      requestId: 'c45dd708-a418-42ec-b8a7-b70a6c6fab0a',
+      bidderRequestId: '178e34bad3658f',
+      bids: [
+        {
+          bidder: 'rubicon',
+          params: {
+            accountId: '14062',
+            siteId: '70608',
+            zoneId: '335918',
+            userId: '12346',
+            keywords: ['a','b','c'],
+            inventory: {
+              rating:'5-star',
+              prodtype:'tech'
+            },
+            visitor: {
+              ucat:'new',
+              lastsearch:'iphone'
+            },
+            position: 'atf',
+            referrer: 'localhost'
+          },
+          placementCode: '/19968336/header-bid-tag-0',
+          sizes: [[300, 250], [320, 50]],
+          bidId: '2ffb201a808da7',
+          bidderRequestId: '178e34bad3658f',
+          requestId: 'c45dd708-a418-42ec-b8a7-b70a6c6fab0a'
+        }
+      ],
+      start: 1472239426002,
+      timeout: 5000
     };
 
   });
@@ -50,15 +84,13 @@ describe("the rubicon adapter", () => {
     sandbox.restore();
   });
 
-  describe("callBids public interface", () => {
+  describe('callBids public interface', () => {
 
-    beforeEach(() => {
-      rubiconAdapter = adapterManager.bidderRegistry["rubicon"];
-    });
+    let rubiconAdapter = adapterManager.bidderRegistry['rubicon'];
 
-    it("should receive a well-formed bidRequest from the adaptermanager", () => {
+    it('should receive a well-formed bidRequest from the adaptermanager', () => {
 
-      sandbox.stub(rubiconAdapter, "callBids");
+      sandbox.stub(rubiconAdapter, 'callBids');
 
       adapterManager.callBids({
           adUnits: [clone(adUnit)]
@@ -66,23 +98,23 @@ describe("the rubicon adapter", () => {
 
       let bidderRequest = rubiconAdapter.callBids.getCall(0).args[0];
 
-      expect(bidderRequest).to.have.property("bids")
-        .that.is.an("array")
+      expect(bidderRequest).to.have.property('bids')
+        .that.is.an('array')
         .with.lengthOf(1);
 
-      expect(bidderRequest).to.have.deep.property("bids[0]")
-        .to.have.property("bidder", "rubicon");
+      expect(bidderRequest).to.have.deep.property('bids[0]')
+        .to.have.property('bidder', 'rubicon');
 
-      expect(bidderRequest).to.have.deep.property("bids[0]")
-        .to.have.property("placementCode", adUnit.code);
+      expect(bidderRequest).to.have.deep.property('bids[0]')
+        .to.have.property('placementCode', adUnit.code);
 
-      expect(bidderRequest).to.have.deep.property("bids[0]")
-        .with.property("sizes")
-        .that.is.an("array")
+      expect(bidderRequest).to.have.deep.property('bids[0]')
+        .with.property('sizes')
+        .that.is.an('array')
         .with.lengthOf(2)
         .that.deep.equals(adUnit.sizes);
 
-      expect(bidderRequest).to.have.deep.property("bids[0]")
+      expect(bidderRequest).to.have.deep.property('bids[0]')
         .with.property('params')
         .that.deep.equals(adUnit.bids[0].params)
 
@@ -90,346 +122,305 @@ describe("the rubicon adapter", () => {
 
   });
 
-  describe("callBids implementation", () => {
+  describe('MAS mapping / ordering', () => {
 
-    let bidderRequest,
-        slot;
+    let masSizeOrdering = RubiconAdapter.masSizeOrdering;
 
-    beforeEach(() => {
-      rubiconAdapter = new RubiconAdapter();
+    it('should not include values without a proper mapping', () => {
+      // two invalid sizes included: [42, 42], [1, 1]
+      let ordering = masSizeOrdering([[320, 50], [42, 42], [300, 250], [640, 480], [1, 1], [336, 280]]);
 
-      sandbox.stub(adloader, "loadScript");
-      sandbox.spy(rubiconAdapter, "callBids");
-
-      slot = {
-        clearTargeting: sandbox.spy(),
-        setPosition: sandbox.spy(),
-        addFPV: sandbox.spy(),
-        addFPI: sandbox.spy(),
-        addKW: sandbox.spy(),
-        getElementId: () => "/19968336/header-bid-tag-0",
-        getRawResponses: () => {},
-        getRawResponseBySizeId: () => {},
-        getAdServerTargetingByKey: key => ({
-          "rpfl_14062": ["15_tier"],
-          "rpfl_elemid": ["/19968336/header-bid-tag-0"]
-        }[key])
-      };
-
-      window.rubicontag = {
-        cmd: {
-          push: cb => cb()
-        },
-        setIntegration: () => {},
-        run: () => {},
-        addEventListener: () => {},
-        setUserKey: sandbox.spy(),
-        defineSlot: sandbox.spy(bid => slot)
-      };
-
-      bidderRequest = {
-        bidderCode: "rubicon",
-        requestId: "c45dd708-a418-42ec-b8a7-b70a6c6fab0a",
-        bidderRequestId: "178e34bad3658f",
-        bids: [
-          {
-            bidder: "rubicon",
-            params: {
-              accountId: "14062",
-              siteId: "70608",
-              zoneId: "335918",
-              userId: "12346",
-              keywords: ["a","b","c"],
-              inventory: {
-                rating:"5-star",
-                prodtype:"tech"
-              },
-              visitor: {
-                ucat:"new",
-                lastsearch:"iphone"
-              },
-              position: "atf"
-            },
-            placementCode: "/19968336/header-bid-tag-0",
-            sizes: [[300, 250], [320, 50]],
-            bidId: "2ffb201a808da7",
-            bidderRequestId: "178e34bad3658f",
-            requestId: "c45dd708-a418-42ec-b8a7-b70a6c6fab0a"
-          }
-        ],
-        start: 1472239426002,
-        timeout: 5000
-      };
-
+      expect(ordering).to.deep.equal([15, 16, 43, 65]);
     });
 
-    it("should load the fastlane SDK if not loaded", () => {
+    it('should sort values without any MAS priority sizes in regular ascending order', () => {
+      let ordering = masSizeOrdering([[320, 50], [640, 480], [336, 280], [200, 600]]);
 
-      rubiconAdapter.callBids(bidderRequest);
-
-      let pathToSDK = adloader.loadScript.getCall(0).args[0];
-      expect(pathToSDK).to.equal(`http://ads.rubiconproject.com/header/${bidderRequest.bids[0].params.accountId}.js`);
-
-      rubiconAdapter.callBids(bidderRequest);
-      expect(adloader.loadScript.calledOnce).to.equal(true);
-
+      expect(ordering).to.deep.equal([16, 43, 65, 126]);
     });
 
-    describe("when doing fastlane slot configuration", () => {
+    it('should sort MAS priority sizes in the proper order w/ rest ascending', () => {
+      let ordering = masSizeOrdering([[320, 50], [160,600], [640, 480], [300, 250],[336, 280], [200, 600]]);
+      expect(ordering).to.deep.equal([15, 9, 16, 43, 65, 126]);
+
+      ordering = masSizeOrdering([[320, 50], [300, 250], [160,600], [640, 480],[336, 280], [200, 600], [728, 90]]);
+      expect(ordering).to.deep.equal([15, 2, 9, 16, 43, 65, 126]);
+
+      ordering = masSizeOrdering([[120, 600], [320, 50], [160,600], [640, 480],[336, 280], [200, 600], [728, 90]]);
+      expect(ordering).to.deep.equal([2, 9, 8, 16, 43, 65, 126]);
+    })
+
+  });
+
+  describe('callBids implementation', () => {
+
+    let rubiconAdapter;
+
+    describe('requests', () => {
+
+      let xhr,
+          screen;
 
       beforeEach(() => {
-        sandbox.spy(window.rubicontag, 'setIntegration');
+        rubiconAdapter = new RubiconAdapter();
+
+        xhr = sandbox.useFakeXMLHttpRequest();
+      });
+
+      afterEach(() => {
+        xhr.restore();
+      });
+
+      it('should make a well-formed optimized request', () => {
 
         rubiconAdapter.callBids(bidderRequest);
-      });
 
-      it("should make a valid call to rubicontag.defineSlot", () => {
+        let request = xhr.requests[0];
 
-        expect(window.rubicontag.defineSlot.calledOnce).to.equal(true);
+        expect(request instanceof sinon.FakeXMLHttpRequest).to.equal(true);
 
-        let slotParam = window.rubicontag.defineSlot.firstCall.args[0];
-        expect(slotParam).to.contain.all.keys(
-          "siteId",
-          "zoneId",
-          "sizes",
-          "id"
+        expect(request.withCredentials).to.equal(true);
+
+        let [path, query] = request.url.split('?');
+        query = parseQuery(query);
+
+        expect(path).to.equal(
+          '//fastlane.rubiconproject.com/a/api/fastlane.json'
         );
-        expect(slotParam).to.have.property("sizes")
-          .that.is.an("array")
-          .with.lengthOf(2)
-          .that.deep.equals([15, 43]);
 
-      });
+        let expectedQuery = {
+          'account_id': '14062',
+          'site_id': '70608',
+          'zone_id': '335918',
+          'size_id': '15',
+          'alt_size_ids': '43',
+          'p_pos': 'atf',
+          'rp_floor': '0.01',
+          'tk_flint': 'pbjs.lite',
+          'p_screen_res': /\d+x\d+/,
+          'tk_user_key': '12346',
+          'kw': 'a,b,c',
+          'tg_v.ucat': 'new',
+          'tg_v.lastsearch': 'iphone',
+          'tg_i.rating': '5-star',
+          'tg_i.prodtype': 'tech',
+          'rf': 'localhost'
+        };
 
-      it("should call rubicontag.setUserKey when params.userId is set", () => {
+        // test that all values above are both present and correct
+        Object.keys(expectedQuery).forEach(key => {
+          let value = expectedQuery[key];
+          if(value instanceof RegExp) {
+            expect(query[key]).to.match(value);
+          } else {
+            expect(query[key]).to.equal(value);
+          }
+        });
 
-        expect(window.rubicontag.setUserKey.calledWith(adUnit.bids[0].params.userId)).to.equal(true);
-
-      });
-
-      it("should set proper targeting params for Slot when passed", () => {
-
-        expect(slot.setPosition.calledOnce).to.equal(true);
-        expect(slot.setPosition.firstCall.calledWith("atf")).to.equal(true);
-
-        expect(slot.addFPV.calledTwice).to.equal(true);
-        expect(slot.addFPV.firstCall.calledWith("ucat", "new")).to.equal(true);
-        expect(slot.addFPV.secondCall.calledWith("lastsearch", "iphone")).to.equal(true);
-
-        expect(slot.addFPI.calledTwice).to.equal(true);
-        expect(slot.addFPI.firstCall.calledWith("rating", "5-star")).to.equal(true);
-        expect(slot.addFPI.secondCall.calledWith("prodtype", "tech")).to.equal(true);
-
-        expect(slot.addKW.calledOnce).to.equal(true);
-        expect(slot.addKW.firstCall.calledWith(["a","b","c"])).to.equal(true);
-
-      });
-
-      it("should set the rubicontag integration as prebid.js", () => {
-
-        expect(window.rubicontag.setIntegration.calledWith("$$PREBID_GLOBAL$$")).to.equal(true);
+        expect(query).to.have.property('rand');
 
       });
 
     });
 
-    describe("rubicon targeting", () => {
+
+    describe('response handler', () => {
+      let bids,
+          server;
 
       beforeEach(() => {
-        // need new adapter to reset private state
-        rubiconAdapter = new RubiconAdapter();
+        bids = [];
+
+        server = sinon.fakeServer.create();
+
+        sandbox.stub(bidManager, 'addBidResponse', (elemId, bid) => {
+          bids.push(bid);
+        });
       });
 
-      it("should register default bidder settings (once) when configured", () => {
+      afterEach(() => {
+        server.restore();
+      });
 
-        sandbox.spy(bidManager, "registerDefaultBidderSetting");
+      it('should handle a success response and sort by cpm', () => {
 
-        sandbox.stub(window.rubicontag, 'setIntegration', () => ({
-          pbjsRubiconTargeting: true
+        server.respondWith(JSON.stringify({
+          "status": "ok",
+          "account_id": 14062,
+          "site_id": 70608,
+          "zone_id": 530022,
+          "size_id": 15,
+          "alt_size_ids": [
+            43
+          ],
+          "tracking": "",
+          "inventory": {},
+          "ads": [
+            {
+              "status": "ok",
+              "impression_id": "153dc240-8229-4604-b8f5-256933b9374c",
+              "size_id": "15",
+              "ad_id": "6",
+              "advertiser": 7,
+              "network": 8,
+              "creative_id": 9,
+              "type": "script",
+              "script": "alert('foo')",
+              "campaign_id": 10,
+              "cpm": 0.811,
+              "targeting": [
+                {
+                  "key": "rpfl_14062",
+                  "values": [
+                    "15_tier_all_test"
+                  ]
+                }
+              ]
+            },
+            {
+              "status": "ok",
+              "impression_id": "153dc240-8229-4604-b8f5-256933b9374d",
+              "size_id": "43",
+              "ad_id": "7",
+              "advertiser": 7,
+              "network": 8,
+              "creative_id": 9,
+              "type": "script",
+              "script": "alert('foo')",
+              "campaign_id": 10,
+              "cpm": 0.911,
+              "targeting": [
+                {
+                  "key": "rpfl_14062",
+                  "values": [
+                    "15_tier_all_test"
+                  ]
+                }
+              ]
+            }
+          ]
         }));
 
         rubiconAdapter.callBids(bidderRequest);
+
+        server.respond();
+
+        expect(bidManager.addBidResponse.calledTwice).to.equal(true);
+
+        expect(bids).to.be.lengthOf(2);
+
+        expect(bids[0].getStatusCode()).to.equal(CONSTANTS.STATUS.GOOD);
+        expect(bids[0].bidderCode).to.equal("rubicon");
+        expect(bids[0].width).to.equal(320);
+        expect(bids[0].height).to.equal(50);
+        expect(bids[0].cpm).to.equal(0.911);
+        expect(bids[0].ad).to.contain(`alert('foo')`)
+          .and.to.contain(`<html>`)
+          .and.to.contain(`<div data-rp-impression-id='153dc240-8229-4604-b8f5-256933b9374d'>`);
+
+        expect(bids[1].getStatusCode()).to.equal(CONSTANTS.STATUS.GOOD);
+        expect(bids[1].bidderCode).to.equal("rubicon");
+        expect(bids[1].width).to.equal(300);
+        expect(bids[1].height).to.equal(250);
+        expect(bids[1].cpm).to.equal(0.811);
+        expect(bids[1].ad).to.contain(`alert('foo')`)
+          .and.to.contain(`<html>`)
+          .and.to.contain(`<div data-rp-impression-id='153dc240-8229-4604-b8f5-256933b9374c'>`);
+      });
+
+      it('should be fine with a CPM of 0', () => {
+        server.respondWith(JSON.stringify({
+          "status": "ok",
+          "account_id": 14062,
+          "site_id": 70608,
+          "zone_id": 530022,
+          "size_id": 15,
+          "alt_size_ids": [
+            43
+          ],
+          "tracking": "",
+          "inventory": {},
+          "ads": [{
+              "status": "ok",
+              "cpm": 0,
+              "size_id": 15
+            }]
+        }));
+
         rubiconAdapter.callBids(bidderRequest);
 
-        expect(bidManager.registerDefaultBidderSetting.calledOnce).to.equal(true);
+        server.respond();
 
+        expect(bidManager.addBidResponse.calledOnce).to.equal(true);
+        expect(bids).to.be.lengthOf(1);
+        expect(bids[0].getStatusCode()).to.equal(CONSTANTS.STATUS.GOOD);
       });
 
-      it("should not register default bidder settings when not configured", () => {
-
-        sandbox.spy(bidManager, "registerDefaultBidderSetting");
+      it('should handle an error with no ads returned', () => {
+        server.respondWith(JSON.stringify({
+          "status": "ok",
+          "account_id": 14062,
+          "site_id": 70608,
+          "zone_id": 530022,
+          "size_id": 15,
+          "alt_size_ids": [
+            43
+          ],
+          "tracking": "",
+          "inventory": {},
+          "ads": []
+        }));
 
         rubiconAdapter.callBids(bidderRequest);
 
-        expect(bidManager.registerDefaultBidderSetting.called).to.equal(false);
+        server.respond();
 
-      })
-
-    });
-
-    describe("when handling fastlane responses", () => {
-
-      describe("individually through events", () => {
-
-        let bids;
-        let _callback;
-        let addEventListener;
-
-        beforeEach(() => {
-          bids = [];
-
-          addEventListener = sandbox.stub(window.rubicontag, "addEventListener", (event, callback) => {
-            _callback = callback;
-            return true;
-          });
-
-          sandbox.stub(bidManager, 'addBidResponse', (elemId, bid) => {
-            bids.push(bid);
-          });
-        });
-
-        it("should only register one listener for multiple bid requests", () => {
-
-          rubiconAdapter.callBids(bidderRequest);
-          rubiconAdapter.callBids(bidderRequest);
-
-          expect(addEventListener.calledOnce).to.equal(true);
-
-        });
-
-        it("should register successful bids with the bidmanager", () => {
-
-          sandbox.stub(window.rubicontag, "run", () => {
-            _callback({
-              elementId: "/19968336/header-bid-tag-0",
-              sizeId: "43"
-            });
-            _callback({
-              elementId: "/19968336/header-bid-tag-0",
-              sizeId: "15"
-            });
-          });
-
-          sandbox.stub(slot, "getRawResponseBySizeId", (sizeId) => {
-            return {
-              "43": {
-                "advertiser": 12345,
-                "cpm": 0.811,
-                "dimensions": [
-                  300,
-                  250
-                ],
-                "auction_id": "431ee1bc-3cc4-4bb7-b0d4-eb9faedb433c"
-              },
-              "15": {
-                "advertiser": 12345,
-                "cpm": 0.59,
-                "dimensions": [
-                  320,
-                  50
-                ],
-                "auction_id": "431ee1bc-3cc4-4bb7-b0d4-eb9faedb433c"
-              }
-            }[sizeId];
-          });
-
-          rubiconAdapter.callBids(bidderRequest);
-
-          expect(bidManager.addBidResponse.calledTwice).to.equal(true);
-
-          expect(bids).to.be.lengthOf(2);
-          expect(bids[0].getStatusCode()).to.equal(CONSTANTS.STATUS.GOOD);
-          expect(bids[1].getStatusCode()).to.equal(CONSTANTS.STATUS.GOOD);
-
-          expect(bids[0].bidderCode).to.equal("rubicon");
-          expect(bids[0].width).to.equal(300);
-          expect(bids[0].height).to.equal(250);
-          expect(bids[0].cpm).to.equal(0.811);
-          expect(bids[0].rubiconTargeting['rpfl_14062']).to.equal("15_tier");
-          expect(bids[0].rubiconTargeting['rpfl_elemid']).to.equal("/19968336/header-bid-tag-0");
-
-          expect(bids[1].bidderCode).to.equal("rubicon");
-          expect(bids[1].width).to.equal(320);
-          expect(bids[1].height).to.equal(50);
-          expect(bids[1].cpm).to.equal(0.59);
-          expect(bids[1].rubiconTargeting['rpfl_14062']).to.equal("15_tier");
-          expect(bids[1].rubiconTargeting['rpfl_elemid']).to.equal("/19968336/header-bid-tag-0");
-
-        })
-
+        expect(bidManager.addBidResponse.calledOnce).to.equal(true);
+        expect(bids).to.be.lengthOf(1);
+        expect(bids[0].getStatusCode()).to.equal(CONSTANTS.STATUS.NO_BID);
       });
 
-      describe("all at once", () => {
+      it('should handle an error with bad status', () => {
+        server.respondWith(JSON.stringify({
+          "status": "ok",
+          "account_id": 14062,
+          "site_id": 70608,
+          "zone_id": 530022,
+          "size_id": 15,
+          "alt_size_ids": [
+            43
+          ],
+          "tracking": "",
+          "inventory": {},
+          "ads": [{
+              "status": "not_ok",
+            }]
+        }));
 
-        let bids;
+        rubiconAdapter.callBids(bidderRequest);
 
-        beforeEach(() => {
-          bids = [];
+        server.respond();
 
-          sandbox.stub(window.rubicontag, "run", cb => cb());
-          sandbox.stub(window.rubicontag, "addEventListener", () => false);
-          sandbox.stub(bidManager, 'addBidResponse', (elemId, bid) => {
-            bids.push(bid);
-          });
-        });
-
-        it("should register successful bids with the bidmanager", () => {
-
-          sandbox.stub(slot, "getRawResponses", () => [
-            {
-              "advertiser": 12345,
-              "cpm": 0.811,
-              "dimensions": [
-                300,
-                250
-              ],
-              "auction_id": "431ee1bc-3cc4-4bb7-b0d4-eb9faedb433c"
-            },
-            {
-              "advertiser": 123456,
-              "cpm": 0.59,
-              "dimensions": [
-                320,
-                50
-              ],
-              "auction_id": "a3e042e5-3fb7-498f-b60e-71540f4769a8"
-            }
-          ]);
-
-          rubiconAdapter.callBids(bidderRequest);
-
-          expect(bidManager.addBidResponse.calledTwice).to.equal(true);
-
-          expect(bids).to.be.lengthOf(2);
-          expect(bids[0].getStatusCode()).to.equal(CONSTANTS.STATUS.GOOD);
-          expect(bids[1].getStatusCode()).to.equal(CONSTANTS.STATUS.GOOD);
-
-          expect(bids[0].bidderCode).to.equal("rubicon");
-          expect(bids[0].width).to.equal(300);
-          expect(bids[0].height).to.equal(250);
-          expect(bids[0].cpm).to.equal(0.811);
-
-          expect(bids[1].bidderCode).to.equal("rubicon");
-          expect(bids[1].width).to.equal(320);
-          expect(bids[1].height).to.equal(50);
-          expect(bids[1].cpm).to.equal(0.59);
-
-        });
-
-        it("should register bad responses as errors with the bidmanager", () => {
-
-          sandbox.stub(slot, "getRawResponses", () => []);
-
-          rubiconAdapter.callBids(bidderRequest);
-
-          expect(bidManager.addBidResponse.calledOnce).to.equal(true);
-          expect(bids[0].getStatusCode()).to.equal(CONSTANTS.STATUS.NO_BID);
-
-        });
-
+        expect(bidManager.addBidResponse.calledOnce).to.equal(true);
+        expect(bids).to.be.lengthOf(1);
+        expect(bids[0].getStatusCode()).to.equal(CONSTANTS.STATUS.NO_BID);
       });
 
-    });
+      it('should handle an error because of malformed json response', () => {
+        server.respondWith("{test{");
+
+        rubiconAdapter.callBids(bidderRequest);
+
+        server.respond();
+
+        expect(bidManager.addBidResponse.calledOnce).to.equal(true);
+        expect(bids).to.be.lengthOf(1);
+        expect(bids[0].getStatusCode()).to.equal(CONSTANTS.STATUS.NO_BID);
+        expect(bids[0].error instanceof SyntaxError).to.equal(true);
+      });
+
+    })
+
 
   });
